# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice template with hexagonal architecture",
        "title": "Monitoring Energy Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "paths": {
...
        "/api/v1/test/intake": {
            "post": {
                "description": "Simulates a Kafka message to test validation and Telegram notifications",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "Test intake message processing",
                "parameters": [
                    {
                        "description": "Test intake data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.TestIntakeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "All validations passed",
                        "schema": {
                            "$ref": "#/definitions/rest.TestIntakeResponse"
                        }
                    },
                    "400": {
                        "description": "Validation failed (invalid UUID or missing field)",
                        "schema": {
                            "$ref": "#/definitions/rest.TestIntakeResponse"
                        }
                    },
                    "404": {
                        "description": "Plant does not exist in database",
                        "schema": {
                            "$ref": "#/definitions/rest.TestIntakeResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.TestIntakeRequest": {
            "type": "object",
            "properties": {
                "event_type": {
                    "type": "string",
                    "example": "power_reading"
                },
                "plant_name": {
                    "type": "string",
                    "example": "planta_solar_1"
                },
                "plant_source_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "rest.TestIntakeResponse": {
            "type": "object",
            "properties": {
                "success": {
                    "type": "boolean",
                    "example": false
                },
                "message": {
                    "type": "string",
                    "example": "Validation failed: invalid UUID format"
                },
                "telegram_sent": {
                    "type": "boolean",
                    "example": true
                },
                "error": {
                    "type": "string",
                    "example": "invalid plant_source_id format"
                }
            }
        },
...
    }
}